+++ src/main/java/cz/cvut/fsv/webgama/dao/PointDao.java
@@ -7,7 +7,7 @@ import cz.cvut.fsv.webgama.domain.Point;
 
 public interface PointDao {
 	
-	public void insert(Point point, Integer networkId);
+	public void insert(Point point, Long networkId);
 
 	public void delete(Point point);
